Add visualisation_pipeline which includes cilPlaneClipper, to clip polydata on 2D viewers.
