+++ Wrappers/Python/ccpi/viewer/utils/__init__.py
@@ -1,4 +1,6 @@
 from .conversion import Converter, cilRegularPointCloudToPolyData, cilMaskPolyData,\
-                        cilClipPolyDataBetweenPlanes, cilNumpyMETAImageWriter,\
+                        cilNumpyMETAImageWriter,\
                         WriteNumpyAsMETAImage
 from .colormaps import *
+
+from .visualisation_pipeline import cilClipPolyDataBetweenPlanes, cilPlaneClipper
